lab_1/chat/utils.py
from jose import JWTError, jwt
from fastapi import HTTPException
from typing import Optional

import httpx

SECRET_KEY = "secret"
ALGORITHM = "HS256"

# Function to decode JWT and get user information
def get_user_from_token(token: str) -> Optional[str]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_email: str = payload.get("sub")
        if user_email is None:
            raise HTTPException(status_code=401, detail="Invalid token")
        return user_email
    except JWTError:
        raise HTTPException(status_code=401, detail="Invalid token")
    
def verify_jwt_token(token):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=["HS256"])
        return payload
    except jwt.JWTError:
        return None