lab_1/chat/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ecdsa==0.19.0
fastapi==0.115.6
flake8==7.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
mccabe==0.7.0
numpy==2.2.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pycodestyle==2.12.1
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-jose==3.3.0
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
Jinja2==3.1.5
MarkupSafe==3.0.2
websockets==14.2
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
